Update mock recordings to make them PXE-bootable
This update ensures that the Redfish mock can be used for testing
physical infrastructure provisioning (network OS boot) workflow
simulation. The EthernetInterfaces entries are necessary for the
PXE to be informed about the MAC addresses of the connected network
interfaces. Even though they are currently all connected to the
same Chassis, this is a minimal change that makes the process
work.
We also had to change the power state of one of the service from
a transitional `PowerOn` to a final `On` state, because the mock
server tool doesn't support moving from this state.